Reverts Keycloak Java libs to 2.4.0.Final
For whatever reason, 2.5.0.Final is not available in Maven
Central yet.  Also, the Keycloak repo only references
2.5.0-Final-SNAPSHOT, which makes me think they only built it
locally, never pushing it up.  Perhaps there we no changes to
the underlying Java libs, so no need to bump the version?  It
is unclear, but I can say that building with 2.4.0.Final does
_seem_ to be working under the 2.5.0.Final server.